Renamed utill to common according to review suggestion
