+++ uikit/src/main/java/care/intouch/uikit/common/ImageVO.kt
@@ -1,4 +1,4 @@
-package care.intouch.uikit.utill
+package care.intouch.uikit.common
 
 import androidx.annotation.DrawableRes
 import androidx.compose.runtime.Composable
